refactor: move opt module into cli
Closes #61
