+++ src/cli/mod.rs
@@ -1,5 +1,7 @@
 //! Command-line interface boundary.
 //!
 //! This module is the target home for argument parsing, command dispatch,
-//! terminal presentation, and exit-code mapping. Existing CLI-owned modules
-//! remain in their legacy paths until a later mechanical move.
+//! terminal presentation, and exit-code mapping. Remaining CLI-owned modules
+//! stay in their legacy paths until later mechanical moves.
+
+pub mod opt;
