+++ src/lib/mod.rs
@@ -8,8 +8,9 @@ pub mod command;
 pub mod common;
 pub mod lockfile;
 pub mod node_linker;
-pub mod opt;
 pub mod package_manifest;
 pub mod parser;
 pub mod registry;
 pub mod util;
+
+pub use cli::opt;
